refactor: move components/PokeDetails to pages/MainLayout
