+++ react-components/src/pages/MainLayout/PokeDetails/PokeDetails.module.scss
